More useless float()

+++ ipi/engine/motion/dynamics.py
@@ -599,7 +599,7 @@ class NVTCCIntegrator(NVTIntegrator):
         # NB we only have to take into account the energy balance of zeroing centroid velocity when we had added energy through the thermostat
         pnm0 = dstrip(self.nm.pnm)[0]
         dynm30 = dstrip(self.nm.dynm3)[0]
-        self.ensemble.eens += float(0.5 * xp.vecdot(pnm0, pnm0 / dynm30))
+        self.ensemble.eens += 0.5 * xp.vecdot(pnm0, pnm0 / dynm30)
         self.nm.pnm[0, :] = 0.0
 
         self.pstep()
@@ -618,7 +618,7 @@ class NVTCCIntegrator(NVTIntegrator):
         self.thermostat.step()
         pnm0 = dstrip(self.nm.pnm)[0]
         dynm30 = dstrip(self.nm.dynm3)[0]
-        self.ensemble.eens += float(0.5 * xp.vecdot(pnm0, pnm0 / dynm30))
+        self.ensemble.eens += 0.5 * xp.vecdot(pnm0, pnm0 / dynm30)
         self.nm.pnm[0, :] = 0.0
         self.pconstraints()
